requirements.txt
asgiref==3.11.1
charset-normalizer==3.4.5
Django==6.0.3
et_xmlfile==2.0.0
openpyxl==3.1.5
pillow==12.1.1
psycopg2-binary==2.9.10
PyMonCtl==0.92
pyinstaller==6.10.0
pywebview==5.3
pywin32==311
PyWinBox==0.7
PyWinCtl==0.4.1
reportlab==4.4.10
sqlparse==0.5.5
typing_extensions==4.15.0
tzdata==2025.3
